Access captures per capture resolution

Captures per capture resolution and fishery group are returned.

This function can used to access capture code resolutions contained within the sefraData object.

Usage

capture_resolutions(object, aggregate = TRUE, ...)

# S4 method for class 'sefraData'
capture_resolutions(object, aggregate)

Arguments

object

a sefraData class object.

aggregate

logical value indicating whether captures should be aggregated by code resolution. Defaults to TRUE. Otherwise captures are returned by fishery group and capture code.

...

additional arguments to the generic function (not used).

Note

Capture code resolutions are listed in code_resolution.

Examples

# assign character vector
dat <- sefraData(c('DIW', 'DQS', 'TWD'))
#> species codes input: including all possible capture codes
#> constructed 'sefraData' object
capture_resolutions(dat)
#> 10 capture code resolutions:
#> (empty captures data frame)
#>    code id_code resolution id_resolution
#> 1   DIW       1    species             1
#> 2   DQS       2    species             2
#> 3   TWD      13    species             3
#> 4   DGA      26    complex             4
#> 5   DST      29    complex             5
#> 6   DWC      32    complex             1
#> 7   DIZ      34      genus             2
#> 8   THZ      35      genus             3
#> 9   ALZ      38     family             4
#> 10  BLZ      40      class             5