Access or assign distribution parameters for \(P^{SH}_{s,m}\)

This function can used to access or assign the distribution specifications for the probability of an adult being in the southern hemisphere per month per species (p_southern) within the sefraData object.

Usage

p_southern(object, ...)

# S4 method for class 'sefraData'
p_southern(object)

p_southern(object, ...) <- value

# S4 method for class 'sefraData,array'
p_southern(object) <- value

Arguments

object

a sefraData class object

...

additional arguments to the generic function (not used)

value

an array containing fixed probability values \(0 \leq x \leq 1\). The array should have twelve columns labelled using abbreviated months. The row names should contain the three-letter species code or codes listed in species.

Note

During assignment of values to an sefraData object, the species codes included in the row names of the array are checked to ensure they match the species contained in the object. If the codes do not match, then the assignment function will return an error.

Examples

library(sefra)
dat <- sefraData(c('DIW', 'DQS', 'TWD'))
#> species codes input: including all species-dependent capture codes
#> constructed 'sefraData' object
p_southern(dat) <- array(runif(36), dim = c(3, 12), 
   dimnames = list(c('DIW', 'DQS', 'TWD'), sefra::months))
p_southern(dat)
#>           Jan       Feb       Mar       Apr        May       Jun        Jul
#> DIW 0.5325652 0.3883498 0.6752085 0.9686375 0.81459620 0.7634748 0.20403065
#> DQS 0.5468048 0.1723519 0.9462948 0.3870963 0.07096477 0.4353866 0.03102602
#> TWD 0.0959265 0.6907258 0.1962195 0.6503439 0.52683032 0.5524723 0.96970706
#>           Aug       Sep        Oct        Nov       Dec
#> DIW 0.1786131 0.8364463 0.03590981 0.69658366 0.1264926
#> DQS 0.7782928 0.6053684 0.13141851 0.40572872 0.9373302
#> TWD 0.8857108 0.9068795 0.09403037 0.06563664 0.2163802